+++ src/ts/Performance/AdUnits/PerformanceAdUnitFactory.ts
@@ -1,20 +1,18 @@
 import { IAdUnitParameters } from 'Ads/AdUnits/AbstractAdUnit';
 import { AbstractAdUnitFactory } from 'Ads/AdUnits/AbstractAdUnitFactory';
 import { IVideoEventHandlerParams } from 'Ads/EventHandlers/BaseVideoEventHandler';
-import { AdUnitStyle } from 'Ads/Models/AdUnitStyle';
-import { IEndScreenParameters } from 'Ads/Views/EndScreen';
 import { PerformanceEndScreenEventHandler } from 'Performance/EventHandlers/PerformanceEndScreenEventHandler';
 import { PerformanceOverlayEventHandler } from 'Performance/EventHandlers/PerformanceOverlayEventHandler';
 import { PerformanceVideoEventHandler } from 'Performance/EventHandlers/PerformanceVideoEventHandler';
 import { PerformanceCampaign } from 'Performance/Models/PerformanceCampaign';
-import { PerformanceEndScreen } from 'Performance/Views/PerformanceEndScreen';
 import { CustomFeatures } from 'Ads/Utilities/CustomFeatures';
 import { Privacy } from 'Ads/Views/Privacy';
 import { Platform } from 'Core/Constants/Platform';
 import { IPerformanceAdUnitParameters, PerformanceAdUnit } from 'Performance/AdUnits/PerformanceAdUnit';
 import { AppStoreDownloadHelper, IAppStoreDownloadHelperParameters } from 'Ads/Utilities/AppStoreDownloadHelper';
+import { AndroidBackButtonSkipTest } from 'Core/Models/ABGroup';
 
-export class PerformanceAdUnitFactory extends AbstractAdUnitFactory {
+export class PerformanceAdUnitFactory extends AbstractAdUnitFactory<PerformanceCampaign, IPerformanceAdUnitParameters> {
 
     public static ContentType = 'comet/campaign';
     public static ContentTypeVideo = 'comet/video';
@@ -24,32 +22,8 @@ export class PerformanceAdUnitFactory extends AbstractAdUnitFactory {
         return contentType === PerformanceAdUnitFactory.ContentType || contentType === PerformanceAdUnitFactory.ContentTypeVideo || contentType === PerformanceAdUnitFactory.ContentTypeMRAID;
     }
 
-    public createAdUnit(parameters: IAdUnitParameters<PerformanceCampaign>): PerformanceAdUnit {
-        const privacy = this.createPrivacy(parameters);
-        const showPrivacyDuringVideo = parameters.placement.skipEndCardOnClose() || false;
-        const overlay = this.createOverlay(parameters, privacy, showPrivacyDuringVideo);
-
-        const adUnitStyle: AdUnitStyle = parameters.campaign.getAdUnitStyle() || AdUnitStyle.getDefaultAdUnitStyle();
-
-        const endScreenParameters: IEndScreenParameters = {
-            ... this.createEndScreenParameters(privacy, parameters.campaign.getGameName(), parameters),
-            adUnitStyle: adUnitStyle,
-            campaignId: parameters.campaign.getId(),
-            osVersion: parameters.deviceInfo.getOsVersion()
-        };
-        const endScreen = new PerformanceEndScreen(endScreenParameters, parameters.campaign, parameters.coreConfig.getCountry());
-        const video = this.getVideo(parameters.campaign, parameters.forceOrientation);
-
-        const performanceAdUnitParameters: IPerformanceAdUnitParameters = {
-            ... parameters,
-            video: video,
-            overlay: overlay,
-            endScreen: endScreen,
-            adUnitStyle: adUnitStyle,
-            privacy: privacy
-        };
-
-        const performanceAdUnit = new PerformanceAdUnit(performanceAdUnitParameters);
+    public createAdUnit(parameters: IPerformanceAdUnitParameters): PerformanceAdUnit {
+        const performanceAdUnit = new PerformanceAdUnit(parameters);
 
         let performanceOverlayEventHandler: PerformanceOverlayEventHandler;
 
@@ -68,19 +42,21 @@ export class PerformanceAdUnitFactory extends AbstractAdUnitFactory {
         };
         const downloadHelper = new AppStoreDownloadHelper(downloadHelperParameters);
 
-        performanceOverlayEventHandler = new PerformanceOverlayEventHandler(performanceAdUnit, performanceAdUnitParameters, downloadHelper);
-        overlay.addEventHandler(performanceOverlayEventHandler);
-        const endScreenEventHandler = new PerformanceEndScreenEventHandler(performanceAdUnit, performanceAdUnitParameters, downloadHelper);
-        endScreen.addEventHandler(endScreenEventHandler);
+        performanceOverlayEventHandler = new PerformanceOverlayEventHandler(performanceAdUnit, parameters, downloadHelper);
+        parameters.overlay.addEventHandler(performanceOverlayEventHandler);
+        const endScreenEventHandler = new PerformanceEndScreenEventHandler(performanceAdUnit, parameters, downloadHelper);
+        parameters.endScreen.addEventHandler(endScreenEventHandler);
 
-        const videoEventHandlerParams = this.getVideoEventHandlerParams(performanceAdUnit, video, performanceAdUnitParameters.adUnitStyle, performanceAdUnitParameters);
+        const videoEventHandlerParams = this.getVideoEventHandlerParams(performanceAdUnit, parameters.video, parameters.adUnitStyle, parameters);
         this.prepareVideoPlayer(PerformanceVideoEventHandler, <IVideoEventHandlerParams<PerformanceAdUnit>>videoEventHandlerParams);
 
         if (parameters.platform === Platform.ANDROID) {
             const onBackKeyObserver = parameters.ads.Android!.AdUnit.onKeyDown.subscribe((keyCode, eventTime, downTime, repeatCount) => {
                 endScreenEventHandler.onKeyEvent(keyCode);
+                const abGroup = parameters.coreConfig.getAbGroup();
+                const backButtonTestEnabled = AndroidBackButtonSkipTest.isValid(abGroup);
 
-                if(CustomFeatures.isCheetahGame(parameters.clientInfo.getGameId())) {
+                if(backButtonTestEnabled || CustomFeatures.isCheetahGame(parameters.clientInfo.getGameId())) {
                     performanceOverlayEventHandler.onKeyEvent(keyCode);
                 }
             });
@@ -90,7 +66,7 @@ export class PerformanceAdUnitFactory extends AbstractAdUnitFactory {
                 }
             });
         }
-        Privacy.setupReportListener(privacy, performanceAdUnit);
+        Privacy.setupReportListener(parameters.privacy, performanceAdUnit);
 
         return performanceAdUnit;
     }