+++ src/ts/VPAID/AdUnits/VPAIDAdUnitFactory.ts
@@ -10,44 +10,30 @@ import { Closer } from 'Ads/Views/Closer';
 import { VPAID } from 'VPAID/Views/VPAID';
 import { Privacy } from 'Ads/Views/Privacy';
 
-export class VPAIDAdUnitFactory extends AbstractAdUnitFactory {
+export class VPAIDAdUnitFactory extends AbstractAdUnitFactory<VPAIDCampaign, IVPAIDAdUnitParameters> {
 
-    public createAdUnit(parameters: IAdUnitParameters<VPAIDCampaign>): VPAIDAdUnit {
-        // WebPlayerContainer will always be defined, checking and throwing just to remove the undefined type.
-        if (!parameters.webPlayerContainer) {
-            throw new Error('is undefined, should not get here.');
+    public createAdUnit(parameters: IVPAIDAdUnitParameters): VPAIDAdUnit {
+        const vpaidAdUnit = new VPAIDAdUnit(parameters);
+
+        const vpaidEventHandler = new VPAIDEventHandler(vpaidAdUnit, {
+            ads: parameters.ads,
+            core: parameters.core,
+            closer: parameters.closer,
+            endScreen: parameters.endScreen,
+            operativeEventManager: parameters.operativeEventManager,
+            thirdPartyEventManager: parameters.thirdPartyEventManager,
+            placement: parameters.placement,
+            campaign: parameters.campaign
+        });
+        parameters.vpaid.addEventHandler(vpaidEventHandler);
+        const overlayEventHandler = new VPAIDOverlayEventHandler(vpaidAdUnit, parameters);
+        parameters.closer.addEventHandler(overlayEventHandler);
+
+        if(parameters.campaign.hasEndScreen() && parameters.endScreen) {
+            const endScreenEventHandler = new VPAIDEndScreenEventHandler(vpaidAdUnit, parameters);
+            parameters.endScreen.addEventHandler(endScreenEventHandler);
         }
-
-        const privacy = this.createPrivacy(parameters);
-        const showGDPRBanner = this.showGDPRBanner(parameters);
-        const closer = new Closer(parameters.platform, parameters.placement, privacy, showGDPRBanner);
-        const vpaid = new VPAID(parameters.platform, parameters.core, parameters.webPlayerContainer, parameters.campaign, parameters.placement);
-        let endScreen: VPAIDEndScreen | undefined;
-
-        const vpaidAdUnitParameters: IVPAIDAdUnitParameters = {
-            ... parameters,
-            vpaid: vpaid,
-            closer: closer,
-            privacy: privacy
-        };
-
-        if(parameters.campaign.hasEndScreen()) {
-            endScreen = new VPAIDEndScreen(parameters.platform, parameters.campaign, parameters.clientInfo.getGameId());
-            vpaidAdUnitParameters.endScreen = endScreen;
-        }
-
-        const vpaidAdUnit = new VPAIDAdUnit(vpaidAdUnitParameters);
-
-        const vpaidEventHandler = new VPAIDEventHandler(vpaidAdUnit, vpaidAdUnitParameters);
-        vpaid.addEventHandler(vpaidEventHandler);
-        const overlayEventHandler = new VPAIDOverlayEventHandler(vpaidAdUnit, vpaidAdUnitParameters);
-        closer.addEventHandler(overlayEventHandler);
-
-        if(parameters.campaign.hasEndScreen() && endScreen) {
-            const endScreenEventHandler = new VPAIDEndScreenEventHandler(vpaidAdUnit, vpaidAdUnitParameters);
-            endScreen.addEventHandler(endScreenEventHandler);
-        }
-        Privacy.setupReportListener(privacy, vpaidAdUnit);
+        Privacy.setupReportListener(parameters.privacy, vpaidAdUnit);
 
         return vpaidAdUnit;
     }